test/requirements.txt
python-docx>=0.8.11
tqdm>=4.62.0 